Add missing controls to the Toolbox
There are many controls that were added since the original Design project was authored.
Some of them were not added due to lack of docs w.r.t the Design project.
So, We'll add them here and now instead.

NOTE: If this works properly, then, we'll backport these changes to older versions.